[homework-7] Second attempt

+++ library_data_jpa/src/main/java/ru/otus/homework7/service/IOService.java
@@ -1,7 +1,7 @@
 package ru.otus.homework7.service;
 
 public interface IOService {
-    long readLong();
+    int readInt();
     String readString();
     void outputString(String text);
 }
